refactor(tui): move parse_date_weekday to shared ui utils
