+++ toki-tui/src/ui/history_panel.rs
@@ -92,7 +92,7 @@ pub fn render_this_week_history(frame: &mut Frame, area: ratatui::layout::Rect, 
                 "── Yesterday ──".to_string()
             } else {
                 // Parse YYYY-MM-DD to get weekday
-                let weekday_label = parse_date_weekday(entry_date);
+                let weekday_label = super::utils::parse_date_weekday(entry_date);
                 format!("── {} ({}) ──", weekday_label, entry_date)
             };
             logical_rows.push(ThisWeekRow::Separator(label));
@@ -237,35 +237,5 @@ pub fn render_this_week_history(frame: &mut Frame, area: ratatui::layout::Rect, 
             inner_area,
             &mut scrollbar_state,
         );
-    }
-}
-
-/// Parse a YYYY-MM-DD string and return the weekday name, or "Unknown" on failure.
-fn parse_date_weekday(date_str: &str) -> &'static str {
-    let parts: Vec<&str> = date_str.splitn(3, '-').collect();
-    if parts.len() != 3 {
-        return "Unknown";
-    }
-    let (Ok(year), Ok(month_u8), Ok(day)) = (
-        parts[0].parse::<i32>(),
-        parts[1].parse::<u8>(),
-        parts[2].parse::<u8>(),
-    ) else {
-        return "Unknown";
-    };
-    let Ok(month) = time::Month::try_from(month_u8) else {
-        return "Unknown";
-    };
-    let Ok(date) = time::Date::from_calendar_date(year, month, day) else {
-        return "Unknown";
-    };
-    match date.weekday() {
-        time::Weekday::Monday => "Monday",
-        time::Weekday::Tuesday => "Tuesday",
-        time::Weekday::Wednesday => "Wednesday",
-        time::Weekday::Thursday => "Thursday",
-        time::Weekday::Friday => "Friday",
-        time::Weekday::Saturday => "Saturday",
-        time::Weekday::Sunday => "Sunday",
     }
 }
